Function opentok_rs_sys::otc_video_frame_new_I420_wrapper

pub unsafe extern "C" fn otc_video_frame_new_I420_wrapper(
    width: c_int, 
    height: c_int, 
    y_plane: *const u8, 
    y_stride: c_int, 
    u_plane: *const u8, 
    u_stride: c_int, 
    v_plane: *const u8, 
    v_stride: c_int
) -> *mut otc_video_frame

Creates a new I420 format video frame from YUV data.

@param width The width of the video, in pixels. @param height The height of the video, in pixels. @param y_plane A pointer the Y plane. @param y_stride The Y stride. @param u_plane A pointer the U plane. @param u_stride The U stride. @param v_plane A pointer the V plane. @param v_stride The V stride. @return A pointer to a video frame instance. This can be null if there is an error.