Struct opentok_rs_sys::otc_audio_device_callbacks

#[repr(C)]
pub struct otc_audio_device_callbacks {

    pub init: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub destroy: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub init_capturer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub destroy_capturer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub start_capturer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub stop_capturer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub is_capturer_initialized: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub is_capturer_started: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub get_estimated_capture_delay: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> c_int>,
    pub get_capture_settings: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void, settings: *mut otc_audio_device_settings) -> otc_bool>,
    pub init_renderer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub destroy_renderer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub start_renderer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub stop_renderer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub is_renderer_initialized: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub is_renderer_started: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>,
    pub get_estimated_render_delay: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> c_int>,
    pub get_render_settings: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void, settings: *mut otc_audio_device_settings) -> otc_bool>,
    pub user_data: *mut c_void,
    pub reserved: *mut c_void,

}

This structure includes a set of function pointers to callback functions to use with an audio device. The SDK calls these functions when events related to the audio device occur.

All callbacks will not be made on the application or main thread but on an internal thread. The application should return the callback as quickly as possible to avoid blocking the internal thread.

In addition to the callbacks, the struct includes a user_data pointer, which points data you can set related to the audio device.

Fields
init: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Called when the SDK requests the audio device to initialize itself.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating either error or success. @see otc_bool

destroy: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Called when the SDK requests the audio device to be destroyed.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating either error or success. @see otc_bool

init_capturer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Called when the SDK requests the audio capturer for the audio device to be initialized.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating either error or success. @see otc_bool

destroy_capturer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Called when the SDK is done capturing audio and it is time to destroy the audio capturer for the audio device.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating either error or success. @see otc_bool

start_capturer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Called when the SDK requests the audio device to start capturing audio samples.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating either error or success. @see otc_bool

stop_capturer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Called when the SDK requests the audio capturer for the audio device to stop sampling audio.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating either error or success. @see otc_bool

is_capturer_initialized: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Whether the device has initialized itself for audio sampling.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating whether the device is initialized or not. @see otc_bool

is_capturer_started: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Whether the device has started audio sampling.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating whether it is started or not. @see otc_bool

get_estimated_capture_delay: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> c_int>

Called when the SDK requests the estimated capturing delay for the audio device, in ms.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return The estimated capturing delay in ms.

get_capture_settings: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void, settings: *mut otc_audio_device_settings) -> otc_bool>

Called when the SDK requests the audio capture settings used by the audio device.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @param settings A pointer to a struct holding the settings. @return Boolean value indicating either error or success. @see otc_bool

init_renderer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Called when the SDK requests the audio device to initialize itself for audio rendering.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating either error or success. @see otc_bool

destroy_renderer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Called when the SDK is done rendering audio and it is time to destroy the audio renderer for the audio device.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating either error or success. @see otc_bool

start_renderer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Called when the SDK requests the audio device to start rendering audio.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating either error or success. @see otc_bool

stop_renderer: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Called when the SDK requests the audio device to stop rendering audio.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating either error or success. @see otc_bool

is_renderer_initialized: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Whether the device has been initialized for audio rendering.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating whether it is initialized or not. @see otc_bool

is_renderer_started: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> otc_bool>

Whether the device started audio rendering.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return Boolean value indicating whether it is started or not. @see otc_bool

get_estimated_render_delay: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void) -> c_int>

Called when the SDK requests the estimated rendering delay for the audio device, in ms.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @return The estimated rendering delay in ms.

get_render_settings: Option<unsafe extern "C" fn(audio_device: *const otc_audio_device, user_data: *mut c_void, settings: *mut otc_audio_device_settings) -> otc_bool>

Called when the SDK requests the audio rendering settings used by the device.

@param audio_device A pointer to the audio device instance. @param user_data A pointer to the user_data you set for the audio device. @param settings A pointer to a struct holding the settings. @return Boolean value indicating either error or success. @see otc_bool

user_data: *mut c_void

A pointer to data you set related to the audio device.

reserved: *mut c_void

A void pointer to a memory area holding reserved resources used for the internal implementation.